🎨 Split chatbot and vector db
가독성과 확장성을 위해 챗봇 기능과 백터 db의 구조를 분리
